# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "contact": {
      "name": "Ploiu",
      "url": "https://github.com/ploiu"
    },
    "title": "Ploiu File Server",
    "description": "a self-hostable file server written in rust",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/password": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createPassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "sets the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "201": {
            "description": "password created"
          },
          "400": {
            "description": "password has already been set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "description": "error occurred setting password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updatePassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "updates the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "204": {
            "description": "password updated"
          },
          "401": {
            "description": "Failed to set password. Check server logs for details"
          }
        }
      }
    },
...
    "/folders/metadata/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "schema": {
            "type": "number",
            "nullable": true,
            "description": "pass null for the root folder; though it cannot be updated or deleted"
          },
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "folders"
        ],
        "description": "retrieves the folder metadata with the passed id",
        "responses": {
          "200": {
            "description": "folder metadata retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/folderMetadata"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "No folder with the passed id was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "folders"
        ],
        "description": "attempts to delete the passed folder. <br><strong>THIS WILL DELETE EVERY FOLDER AND FILE WITHIN THAT FOLDER. THERE IS NO RECOVERING FROM THIS. USE WITH CAUTION</strong>",
        "responses": {
          "204": {
            "description": "folder successfully deleted"
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "That folder could not be found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
    "/folders": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createFolderRequest"
              }
            }
          }
        },
        "description": "creates a folder in the specified location",
        "tags": [
          "folders"
        ],
        "responses": {
          "200": {
            "description": "folder successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/folderMetadata"
                }
              }
            }
          },
          "400": {
            "description": "folder with that name already exists in that location",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "parent folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "folders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createFolderRequest"
              }
            }
          }
        },
        "description": "updates the folder with the passed id",
        "responses": {
          "200": {
            "description": "folder updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/folderMetadata"
                }
              }
            }
          },
          "400": {
            "description": "Action Not Allowed.<br />Either you tried to move a folder into one of its child folders, or you tried to move it to a location already containing a folder with that name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "either the target folder was not found, or the parent folder it's being moved to is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
    "/folders/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "schema": {
            "type": "number",
            "minimum": 0
          },
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "folders"
        ],
        "description": "downloads the folder with the passed id as a tar file. Downloading the root folder will reject the request with a 400. The file server should not be used to back up its own data. Perform backups yourself external to this api",
        "responses": {
          "200": {
            "description": "folder successfully compressed and downloaded",
            "content": {
              "application/x-tar": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "No folder with the passed id was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/genericErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
    "/folders/preview/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "schema": {
            "type": "number"
          },
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "folders"
        ],
        "description": "downloads all file previews under the folder with the passed id",
        "responses": {
          "200": {
            "description": "folder found and previews returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string",
                    "format": "binary"
                  },
                  "description": "map of key => file id : byte[]"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "localhost",
      "url": "https://localhost:8000"
    }
  ],
  "components": {
    "schemas": {
...
      "fileApi": {
        "properties": {
          "id": {
            "type": "number",
            "nullable": false,
            "minimum": 0
          },
          "folderId": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": false
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/tagApi"
            }
          },
          "size": {
            "type": "number",
            "nullable": true,
            "description": "never null in responses, can be null in requests (not modifiable)"
          },
          "dateCreated": {
            "type": "string",
            "nullable": true,
            "description": "never null in responses, can be null in requests (not modifiable)",
            "format": "yyyy-MM-dd'T'HH:mm:ss"
          },
          "fileType": {
            "type": "string",
            "nullable": true,
            "description": "never null in responses, can be null in requests (not modifiable)"
          }
        }
      },
      "createFolderRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "tags": {
            "nullable": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/tagApi"
            }
          }
        }
      },
...
      "basicMessage": {
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "folderMetadata": {
        "properties": {
          "id": {
            "type": "number"
          },
          "parentId": {
            "type": "number",
            "nullable": true
          },
          "path": {
            "type": "string",
            "example": "./root/test"
          },
          "name": {
            "type": "string",
            "example": "test"
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/folderMetadata"
            }
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/fileApi"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/tagApi"
            }
          }
        }
      },
      "tagApi": {
        "properties": {
          "id": {
            "nullable": true,
            "type": "number",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "createPassword": {
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "updatePassword": {
        "properties": {
          "oldAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          },
          "newAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          }
        }
      }
    },
    "responses": {
      "unauthorizedResponse": {
        "description": "missing or invalid password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
      "genericErrorResponse": {
        "description": "something failed on the db or file system side of the operation. Check server logs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}